+++ provider/provider.go
@@ -25,6 +25,7 @@ import (
 	"github.com/databricks/terraform-provider-databricks/clusters"
 	"github.com/databricks/terraform-provider-databricks/commands"
 	"github.com/databricks/terraform-provider-databricks/common"
+	"github.com/databricks/terraform-provider-databricks/dashboards"
 	"github.com/databricks/terraform-provider-databricks/jobs"
 	tflogger "github.com/databricks/terraform-provider-databricks/logger"
 	"github.com/databricks/terraform-provider-databricks/mlflow"
@@ -100,6 +101,7 @@ func DatabricksProvider() *schema.Provider {
 			"databricks_notebook":                             workspace.DataSourceNotebook().ToResource(),
 			"databricks_notebook_paths":                       workspace.DataSourceNotebookPaths().ToResource(),
 			"databricks_pipelines":                            pipelines.DataSourcePipelines().ToResource(),
+			"databricks_schema":                               catalog.DataSourceSchema().ToResource(),
 			"databricks_schemas":                              catalog.DataSourceSchemas().ToResource(),
 			"databricks_service_principal":                    scim.DataSourceServicePrincipal().ToResource(),
 			"databricks_service_principals":                   scim.DataSourceServicePrincipals().ToResource(),
@@ -113,6 +115,7 @@ func DatabricksProvider() *schema.Provider {
 			"databricks_table":                                catalog.DataSourceTable().ToResource(),
 			"databricks_tables":                               catalog.DataSourceTables().ToResource(),
 			"databricks_views":                                catalog.DataSourceViews().ToResource(),
+			"databricks_volume":                               catalog.DataSourceVolume().ToResource(),
 			"databricks_volumes":                              catalog.DataSourceVolumes().ToResource(),
 			"databricks_user":                                 scim.DataSourceUser().ToResource(),
 			"databricks_zones":                                clusters.DataSourceClusterZones().ToResource(),
@@ -129,6 +132,7 @@ func DatabricksProvider() *schema.Provider {
 			"databricks_connection":                      catalog.ResourceConnection().ToResource(),
 			"databricks_cluster":                         clusters.ResourceCluster().ToResource(),
 			"databricks_cluster_policy":                  policies.ResourceClusterPolicy().ToResource(),
+			"databricks_dashboard":                       dashboards.ResourceDashboard().ToResource(),
 			"databricks_dbfs_file":                       storage.ResourceDbfsFile().ToResource(),
 			"databricks_directory":                       workspace.ResourceDirectory().ToResource(),
 			"databricks_entitlements":                    scim.ResourceEntitlements().ToResource(),
@@ -206,6 +210,7 @@ func DatabricksProvider() *schema.Provider {
 			"databricks_vector_search_endpoint":          vectorsearch.ResourceVectorSearchEndpoint().ToResource(),
 			"databricks_vector_search_index":             vectorsearch.ResourceVectorSearchIndex().ToResource(),
 			"databricks_volume":                          catalog.ResourceVolume().ToResource(),
+			"databricks_workspace_binding":               catalog.ResourceWorkspaceBinding().ToResource(),
 			"databricks_workspace_conf":                  workspace.ResourceWorkspaceConf().ToResource(),
 			"databricks_workspace_file":                  workspace.ResourceWorkspaceFile().ToResource(),
 		},
@@ -219,7 +224,7 @@ func DatabricksProvider() *schema.Provider {
 			useragent.WithUserAgentExtra("terraform", p.TerraformVersion)
 		}
 		tflogger.SetLogger()
-		return configureDatabricksClient(ctx, d)
+		return ConfigureDatabricksClient(ctx, d)
 	}
 	common.AddContextToAllResources(p, "databricks")
 	return p
@@ -250,7 +255,7 @@ func providerSchema() map[string]*schema.Schema {
 	return ps
 }
 
-func configureDatabricksClient(ctx context.Context, d *schema.ResourceData) (any, diag.Diagnostics) {
+func ConfigureDatabricksClient(ctx context.Context, d *schema.ResourceData) (any, diag.Diagnostics) {
 	cfg := &config.Config{}
 	attrsUsed := []string{}
 	authsUsed := map[string]bool{}